requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
websockets==13.1
httpx==0.27.2
aiohttp==3.10.10
beautifulsoup4==4.12.3
lxml==5.3.0
selenium==4.25.0
playwright==1.48.0
sqlalchemy==2.0.36
aiosqlite==0.20.0
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1
apscheduler==3.10.4
asyncio-mqtt==0.16.2
aiosmtplib==3.0.2
jinja2==3.1.4
python-multipart==0.0.12
tenacity==9.0.0
redis==5.2.0
celery==5.4.0
loguru==0.7.2
pytest==8.3.3
pytest-asyncio==0.24.0
